used jwt for update and delete blog

+++ backend/db/repository/blog.py
@@ -17,7 +17,8 @@ def retreive_blog(id: int, db: Session):
 
 
 def list_blog(db: Session):
-    blogs = db.query(Blog).filter(Blog.is_active == True).all()
+    # blogs = db.query(Blog).filter(Blog.is_active == True).all()
+    blogs=db.query(Blog).all()
     return blogs
 
 
@@ -25,6 +26,8 @@ def update_blog(id: int, blog: UpdateBlog, author_id: int, db: Session):
     blog_in_db = db.query(Blog).filter(Blog.id == id).first()
     if not blog_in_db:
         return
+    if not blog_in_db.author_id == author_id:
+        return {"error": f"Only the author can modify the blog"}
     blog_in_db.title = blog.title
     blog_in_db.content = blog.content
     db.add(blog_in_db)
@@ -37,6 +40,8 @@ def delete_blog(id: int, author_id: int, db: Session):
     blog_in_db = db.query(Blog).filter(Blog.id == id)
     if not blog_in_db:
         return
+    if not blog_in_db.first().author_id == author_id:
+        return {"error": f"Only the author can delete a blog"}
 
     blog_in_db.delete()
     db.commit()
